Integrated FMOD
Integrated FMOD into the project, including adding the FMOD project to the project folder. I did not disable the built-in unity audio on this project, as I did not know what audio may or may not have already existed.

I tested everything on my end, but please let me know if there is anything broken that I did not catch!